requirements.txt
flask==3.0.0
boto3==1.34.34
python-pptx==0.6.23
msal==1.26.0
markdown==3.5.1
python-dotenv==1.0.0
werkzeug==3.0.1
pdfplumber==0.11.9
